Fixing the reload methods (with proper tearRate) Moving up shoot methods from Hero to LivingCreature class. Moving up List<Projectiles> same Now All living Creature can shoot tear if permite in their proper class.
