+++ src/resources/RoomInfos.java
@@ -11,7 +11,7 @@ public class RoomInfos
 	public static final Vector2 HALF_TILE_SIZE = new Vector2(TILE_WIDTH, TILE_HEIGHT).scalarMultiplication(0.5);
 	
 	public static final Vector2 POSITION_CENTER_OF_ROOM = new Vector2(0.5, 0.5);
-	//TODO MAKE it understandable 0->for wall above and below, 1->mean wall left and right
+	//TODO MAKE it understandable 0->
 	public static final double[] WALL_UP = {0.15*DisplaySettings.SCALE, 1*DisplaySettings.SCALE};
 	public static final double[] WALL_DOWN = {0.21*DisplaySettings.SCALE, 1*DisplaySettings.SCALE};
 	public static final double[] WALL_LEFTnRIGHT = {1*DisplaySettings.SCALE, 0.13*DisplaySettings.SCALE};
